Refactors asteroids game to use hooks
