--- src/components/AsteroidsGame/constants.ts
@@ -1,11 +1,0 @@
-export const ASTEROID_LARGE_POINTS = 20;
-export const ASTEROID_MEDIUM_POINTS = 50;
-export const ASTEROID_SIZE: number = 100; // Starting size of asteroids in pixels.
-export const ASTEROID_SMALL_POINTS = 100;
-export const FONT_SIZE: number = 40; // Font height in pixels.
-export const FONT_TYPE: string = '"VT323", monospace';
-export const FPS: number = 30; // Frames per second.
-export const FRICTION: number = 0.7; // Friction coefficient of space (0 = no friction, 1 = lots of friction).
-export const HIGH_SCORE_KEY: string = '99942-apophis';
-export const SHIP_BLINK_DURATION: number = 0.1; // Duration in seconds of a single blink during ship's invisibility.
-export const SHIP_SIZE: number = 30; // Ship height in pixels.
