+++ src/components/AsteroidsGame/constants/GameConstants.ts
@@ -1,0 +1,4 @@
+export const FONT_SIZE: number = 40; // Font height in pixels.
+export const FONT_TYPE: string = '"VT323", monospace';
+export const FPS: number = 30; // Frames per second.
+export const HIGH_SCORE_KEY: string = '99942-apophis';
