+++ src/components/AsteroidsGame/constants/ShipConstants.ts
@@ -1,0 +1,2 @@
+export const BLINK_DURATION: number = 0.1; // Duration in seconds of a single blink during ship's invisibility.
+export const SIZE: number = 30; // Ship height in pixels.
