+++ src/components/AsteroidsGame/constants/SpaceConstants.ts
@@ -1,0 +1,1 @@
+export const FRICTION: number = 0.7; // Friction coefficient of space (0 = no friction, 1 = lots of friction).
