+++ src/components/AsteroidsGame/constants/index.ts
@@ -1,0 +1,4 @@
+export * as AsteroidConstants from './AsteroidConstants';
+export * as GameConstants from './GameConstants';
+export * as ShipConstants from './ShipConstants';
+export * as SpaceConstants from './SpaceConstants';
