+++ src/components/AsteroidsGame/models/index.ts
@@ -1,0 +1,4 @@
+export { default as Asteroid } from './Asteroid';
+export { default as Explosion } from './Explosion';
+export { default as Particle } from './Particle';
+export { default as Ship } from './Ship';
