+++ src/components/AsteroidsGame/types/ILaser.ts
@@ -1,0 +1,10 @@
+interface ILaser {
+  x: number;
+  y: number;
+  xv: number;
+  yv: number;
+  dist: number;
+  explodeTime: number;
+}
+
+export default ILaser;
