+++ src/components/AsteroidsGame/types/IThrust.ts
@@ -1,0 +1,6 @@
+interface IThrust {
+  x: number;
+  y: number;
+}
+
+export default IThrust;
