+++ src/components/AsteroidsGame/types/index.ts
@@ -1,0 +1,3 @@
+export { default as IEntity } from './IEntity';
+export { default as ILaser } from './ILaser';
+export { default as IThrust } from './IThrust';
