+++ src/components/AsteroidsGame/utils/distanceBetweenPoints.ts
@@ -1,0 +1,8 @@
+export default function distanceBetweenPoints(
+  x1: number,
+  y1: number,
+  x2: number,
+  y2: number
+): number {
+  return Math.sqrt(Math.pow(x2 - x1, 2) + Math.pow(y2 - y1, 2));
+}
